black formatting - pep8

+++ pynsee/macrodata/get_dataset_list.py
@@ -30,14 +30,15 @@ def get_dataset_list():
         INSEE_api_link_dataflow = "https://api.insee.fr/series/BDM/V1/dataflow/FR1/all"
 
         results = _request_insee(
-            api_url=INSEE_api_link_dataflow, sdmx_url=INSEE_sdmx_link_dataflow)
+            api_url=INSEE_api_link_dataflow, sdmx_url=INSEE_sdmx_link_dataflow
+        )
 
         # create temporary directory
         dirpath = _get_temp_dir()
 
-        dataflow_file = dirpath + '\\dataflow_file'
+        dataflow_file = dirpath + "\\dataflow_file"
 
-        with open(dataflow_file, 'wb') as f:
+        with open(dataflow_file, "wb") as f:
             f.write(results.content)
 
         root = ET.parse(dataflow_file).getroot()
@@ -53,15 +54,17 @@ def get_dataset_list():
 
         for i in trange(n_dataflow, desc="Getting datasets list"):
 
-            dataset = {'id': [next(iter(data[i].attrib.values()))],
-                    'Name.fr': [data[i][1].text],
-                    'Name.en': [data[i][2].text],
-                    'url': [data[i][0][0][0].text],
-                    'n_series': [data[i][0][1][0].text]
-                    }
+            dataset = {
+                "id": [next(iter(data[i].attrib.values()))],
+                "Name.fr": [data[i][1].text],
+                "Name.en": [data[i][2].text],
+                "url": [data[i][0][0][0].text],
+                "n_series": [data[i][0][1][0].text],
+            }
 
-            dt = pd.DataFrame(dataset, columns=['id', 'Name.fr', 'Name.en',
-                                                'url', 'n_series'])
+            dt = pd.DataFrame(
+                dataset, columns=["id", "Name.fr", "Name.en", "url", "n_series"]
+            )
             list_df.append(dt)
 
         # concatenate list of dataframes
@@ -70,16 +73,16 @@ def get_dataset_list():
         # clean up columns
         df = df.astype(str)
 
-        df['n_series'] = df['n_series'].str.replace('\\D', '', regex=True)
-        df['n_series'] = df['n_series'].astype('int')
+        df["n_series"] = df["n_series"].str.replace("\\D", "", regex=True)
+        df["n_series"] = df["n_series"].astype("int")
 
         df = df[df["id"] != "SERIES_BDM"]
 
-        df["Name.en"] = df["Name.en"].str.replace('^\\n\\s{0,}', '', regex=True)
-        df["Name.fr"] = df["Name.fr"].str.replace('^\\n\\s{0,}', '', regex=True)
+        df["Name.en"] = df["Name.en"].str.replace("^\\n\\s{0,}", "", regex=True)
+        df["Name.fr"] = df["Name.fr"].str.replace("^\\n\\s{0,}", "", regex=True)
         df = df[df["Name.en"] != ""]
         df = df[df["Name.fr"] != ""]
-    
+
     except:
         df = _get_dataset_list_internal()
